Only assume a free nucleon for phase space kPSQ2fE
Change the reweight calculators to only AssumeFreeNucleon in case the
phase space is kPSQ2fE, since the assumption is no longer generally
valid in Generator v3.00.04, and this leads to cross section mismatches.

Ideally we would use a model identifier to set this flag, but since
it's not readily available we use the phase space stored in the
event record as a proxy.